requirements.txt
python-dotenv==1.0.0
openai==1.2.0
requests==2.31.0
bs4==4.13.0
os==1.0.0